Add more docs, move a shared function out, and properly flip between v1 and v2 on reload
