+++ pkg/clients/cloudwatch/input.go
@@ -1,0 +1,33 @@
+package cloudwatch
+
+import "time"
+
+const TimeFormat = "2006-01-02T15:04:05.999999-07:00"
+
+// Clock small interface which allows for stubbing the time.Now() function for unit testing
+type Clock interface {
+	Now() time.Time
+}
+
+// TimeClock implementation of Clock interface which delegates to Go's Time package
+type TimeClock struct{}
+
+func (tc TimeClock) Now() time.Time {
+	return time.Now()
+}
+
+// determineGetMetricDataWindow computes the start and end time for the GetMetricData request to AWS
+// Always uses the wall clock time as starting point for calculations to ensure that
+// a variety of exporter configurations will work reliably.
+func DetermineGetMetricDataWindow(clock Clock, roundingPeriod time.Duration, length time.Duration, delay time.Duration) (time.Time, time.Time) {
+	now := clock.Now()
+	if roundingPeriod > 0 {
+		// Round down the time to a factor of the period - rounding is recommended by AWS:
+		// https://docs.aws.amazon.com/AmazonCloudWatch/latest/APIReference/API_GetMetricData.html#API_GetMetricData_RequestParameters
+		now = now.Add(-roundingPeriod / 2).Round(roundingPeriod)
+	}
+
+	startTime := now.Add(-(length + delay))
+	endTime := now.Add(-delay)
+	return startTime, endTime
+}
